+++ basalt/src/app.rs
@@ -803,6 +803,7 @@ impl<'a> App<'a> {
                     );
                     editor.set_vim_mode(config.vim_mode);
                     editor.set_editor_enabled(config.experimental_editor);
+                    editor.set_wrap(config.wrap);
                     if config.experimental_editor && config.vim_mode {
                         editor.set_view(View::Edit(EditMode::Source));
                     } else {