+++ basalt/src/note_editor/cursor.rs
@@ -1,17 +1,8 @@
 use std::ops::ControlFlow;
 
-use ratatui::{
-    buffer::Buffer,
-    layout::{Offset, Rect},
-    style::{Color, Style},
-    widgets::StatefulWidget,
-};
 use unicode_width::UnicodeWidthChar;
 
-use crate::{
-    config::Theme,
-    note_editor::{text_buffer::TextBuffer, virtual_document::VirtualLine},
-};
+use crate::note_editor::{text_buffer::TextBuffer, virtual_document::VirtualLine};
 
 #[derive(Clone, Debug)]
 pub enum Message {
@@ -386,63 +377,6 @@ impl Cursor {
 
     pub fn virtual_column(&self) -> usize {
         self.virtual_column
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct CursorWidget {
-    offset: Offset,
-    meta_len: u16,
-    selection: Color,
-}
-
-impl Default for CursorWidget {
-    fn default() -> Self {
-        Self {
-            offset: Offset::default(),
-            meta_len: 0,
-            selection: Theme::default().muted,
-        }
-    }
-}
-
-impl CursorWidget {
-    pub fn with_offset(mut self, offset: Offset) -> Self {
-        self.offset = offset;
-        self
-    }
-
-    pub fn with_meta_len(mut self, meta_len: u16) -> Self {
-        self.meta_len = meta_len;
-        self
-    }
-
-    pub fn with_theme(mut self, theme: &Theme) -> Self {
-        self.selection = theme.muted;
-        self
-    }
-}
-
-impl StatefulWidget for CursorWidget {
-    type State = Cursor;
-
-    fn render(self, area: Rect, buf: &mut Buffer, state: &mut Self::State)
-    where
-        Self: Sized,
-    {
-        if !matches!(state.mode, CursorMode::Read) {
-            return;
-        }
-
-        let y = (state.virtual_row as u16)
-            .saturating_add(self.meta_len)
-            .saturating_sub(area.top())
-            .saturating_add(self.offset.y as u16);
-
-        buf.set_style(
-            Rect::new(self.offset.x as u16, y, area.width, 1),
-            Style::default().reversed().fg(self.selection),
-        );
     }
 }